Planetのサイズ計算にDesimalを使用することでピッタリのサイズで処理が走るようになりました。しかし、地球までの距離にズレが生じています。

+++ Assets/Script/planet/earth.cs
@@ -5,11 +5,11 @@ using UnityEngine;
 public class earth : earth_Distans
 {
     [SerializeField]int earthTime;
-    [SerializeField]decimal earthSpeed;
+    [SerializeField]float earthSpeed;
 
     void Start()
     {
         StartCoroutine(planet_Move(earthTime,earthSpeed));
-        StartCoroutine(earthDistansCount(earthTime * 10,earthSpeed * 0.1m));
+        StartCoroutine(earthDistansCount(earthTime * 10,earthSpeed * 0.1f));
     }
 }
